# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ThoorigAI Infotech Admin Dashboard API
  version: 0.1.0
  description: >-
    App Router HTTP contract. Protected routes require the Supabase cookie session
    and the role permission documented in docs/architecture/role-permissions.md.
    Route handlers and Zod schemas remain authoritative when this reference changes.
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:3000
security:
  - cookieAuth: []
paths:
...
  /api/attendance:
    get:
      operationId: listAttendance
      summary: List course attendance and report data
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/Search' }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/ListResult' }
    post:
      operationId: markAttendance
      summary: Mark or update an enrolled student's attendance
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: Validated by the attendance Zod schema }
      responses:
        '200': { description: Attendance saved }
        '409': { description: Duplicate session or attendance conflict }
...
  /api/assessments/{id}/results:
    get:
      operationId: listAssessmentResults
      summary: List results for an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/ListResult' }
    post:
      operationId: saveAssessmentResults
      summary: Add or update one or more assessment results
      responses:
        '200': { description: Assessment results saved }
    delete:
      operationId: deleteAssessmentResult
      summary: Delete an assessment result
      responses:
        '200': { description: Result deleted }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-<project-ref>-auth-token
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Search:
      name: search
      in: query
      schema: { type: string, maxLength: 100 }
  responses:
    ListResult:
      description: Paginated `{ data, meta }` response
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { type: array, items: { type: object } }
              meta: { type: object }
...